Completely restructured Visualiser components
