--- client/src/components/Visualiser/CodeSnippet/index.ts
@@ -1,1 +1,0 @@
-export { default as CodeSnippet } from './CodeSnippet';
