--- client/src/components/Visualiser/Controller/index.ts
@@ -1,1 +1,0 @@
-export { default as VisualiserControls } from './VisualiserControls';
