--- client/src/components/Visualiser/FloatingWindow/index.ts
@@ -1,1 +1,0 @@
-export { default as FloatingWindow } from './FloatingWindow';
