+++ client/src/components/Visualiser/VisualiserInterface/Controls.tsx
@@ -23,7 +23,7 @@ import {
 import { styled } from '@mui/system';
 import React, { EventHandler, FC, useCallback, useContext, useEffect, useState } from 'react';
 import { defaultSpeed } from 'visualiser-src/common/constants';
-import VisualiserContext from '../VisualiserContext';
+import VisualiserContext from './VisualiserContext';
 import styles from './Control.module.scss';
 
 const TimelineSlider = styled('input')({
