+++ client/src/components/Visualiser/index.ts
@@ -1,2 +1,3 @@
-export { default as VisualiserCanvas } from './VisualiserCanvas';
-export { default as VisualiserInterface } from './VisualiserInterface';
+import Visualiser from './Visualiser';
+
+export default Visualiser;
